requirements.txt
asttokens==2.0.5
backcall==0.2.0
bcrypt==3.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
decorator==5.1.1
dlib==19.23.1
executing==0.8.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.28.5
graphviz==0.19.1
idna==3.3
ipython==8.0.1
jedi==0.18.1
kaggle==1.5.12
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.1
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.0
paramiko==2.9.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
polars==0.12.15
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
Pygments==2.11.2
PyGObject==3.42.0
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
stack-data==0.1.4
text-unidecode==1.3
timm==0.5.4
tk==0.1.0
tomli==1.2.3
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchvision==0.11.2+cu113
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
